backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic
python-dotenv
alembic
